add member and edited member information page and removed unused component
