--- src/pages/HomePage/pages/Members/pages/MembersAssets.jsx
@@ -1,8 +1,0 @@
-import UnderConstruction from "/src/pages/UnderConstruction/UnderConstruction";
-const MembersAssets = () => {
-    return (
-        <UnderConstruction/>
-    );
-}
-
-export default MembersAssets;
