+++ src/routes/appRoutes.jsx
@@ -9,10 +9,21 @@ import EventRegister from "@/pages/HomePage/pages//EventsManagement/pages/EventR
 import ManageAsset from "@/pages/HomePage/pages/AssetsManagement/pages/ManageAssets";
 import EventsManagement from "@/pages/HomePage/pages/EventsManagement/EventsManagement";
 import ViewEvent from "@/pages/HomePage/pages/EventsManagement/pages/ViewEvents.jsx";
-import Requisitions from "@/pages/HomePage/pages/Requisitions/pages/Requisitions";
+import MinistrySchool from "@/pages/HomePage/pages/MinistrySchool/MinistrySchool";
+import ViewCertificate from "@/pages/HomePage/pages/MinistrySchool/pages/ViewCertificate";
+import ViewClass from "@/pages/HomePage/pages/MinistrySchool/pages/ViewClass";
+import ViewCohort from "@/pages/HomePage/pages/MinistrySchool/pages/ViewCohort";
+import ViewProgram from "@/pages/HomePage/pages/MinistrySchool/pages/ViewProgram";
+import ViewStudent from "@/pages/HomePage/pages/MinistrySchool/pages/ViewStudent";
+import MyRequisitions from "@/pages/HomePage/pages/Requisitions/pages/MyRequests";
 import Request from "@/pages/HomePage/pages/Requisitions/pages/Request.tsx";
 import RequestDetails from "@/pages/HomePage/pages/Requisitions/pages/RequestDetails.tsx";
+import Requisitions from "@/pages/HomePage/pages/Requisitions/pages/Requisitions";
+import AccessRights from "@/pages/HomePage/pages/Settings/pages/AccessRights";
+import CreateAccess from "@/pages/HomePage/pages/Settings/pages/CreateAccess";
+import ViewUser from "@/pages/HomePage/pages/Users/pages/ViewUser";
 import UserManagement from "@/pages/HomePage/pages/Users/UserManagement";
+import Registration from "@/pages/Registration/Registration";
 import HomePage from "../pages/HomePage/HomePage";
 import AssetManagement from "../pages/HomePage/pages/AssetsManagement/AssetManagement";
 import DashBoard from "../pages/HomePage/pages/DashBoard/DashBoard";
@@ -21,18 +32,6 @@ import AddMember from "../pages/HomePage/pages/Members/pages/AddMember";
 import Settings from "../pages/HomePage/pages/Settings/Settings.jsx";
 import UnderConstruction from "../pages/UnderConstruction/UnderConstruction.jsx";
 import MemberInformation from "/src/pages/HomePage/pages/Members/pages/MemberInformation";
-import MembersAssets from "/src/pages/HomePage/pages/Members/pages/MembersAssets.jsx";
-import AccessRights from "@/pages/HomePage/pages/Settings/pages/AccessRights";
-import CreateAccess from "@/pages/HomePage/pages/Settings/pages/CreateAccess";
-import ViewUser from "@/pages/HomePage/pages/Users/pages/ViewUser";
-import MyRequisitions from "@/pages/HomePage/pages/Requisitions/pages/MyRequests";
-import MinistrySchool from "@/pages/HomePage/pages/MinistrySchool/MinistrySchool";
-import ViewProgram from "@/pages/HomePage/pages/MinistrySchool/pages/ViewProgram";
-import ViewCohort from "@/pages/HomePage/pages/MinistrySchool/pages/ViewCohort";
-import ViewClass from "@/pages/HomePage/pages/MinistrySchool/pages/ViewClass";
-import ViewStudent from "@/pages/HomePage/pages/MinistrySchool/pages/ViewStudent";
-import ViewCertificate from "@/pages/HomePage/pages/MinistrySchool/pages/ViewCertificate";
-import Registration from "@/pages/Registration/Registration";
 
 
 export const routes = [
@@ -104,12 +103,6 @@ export const routes = [
             path: "info",
             name: "Info",
             element: <MemberInformation />,
-            isPrivate: true,
-            permissionNeeded: "view_members",
-          },
-          {
-            path: "assets",
-            element: <MembersAssets />,
             isPrivate: true,
             permissionNeeded: "view_members",
           },
@@ -215,7 +208,7 @@ export const routes = [
         isPrivate: true,
         permissionNeeded: "view_events",
       },
-      
+
       {
         path: "ministry-school",
         name: "School of Ministry",
